renamed activity package
