+++ app/build.gradle
@@ -27,9 +27,9 @@ dependencies {
     })
     compile 'com.android.support:appcompat-v7:24.2.1'
     compile 'com.android.support:support-v4:24.2.1'
-    compile 'com.android.support:cardview-v7:24.0.+'
-    compile 'com.android.support:recyclerview-v7:24.0.+'
-    compile 'com.android.support:design:24.0.+'
+    compile 'com.android.support:cardview-v7:24.2.1'
+    compile 'com.android.support:recyclerview-v7:24.2.1'
+    compile 'com.android.support:design:24.2.1'
     compile 'com.github.rahatarmanahmed:circularprogressview:2.5.0'
     compile project(':usbSerialForAndroid')
     compile 'com.nshmura:snappysmoothscroller:1.0.0'
