+++ app/src/main/java/com/freshollie/monkeyboard/keystoneradio/ui/RadioStatusDialog.java
@@ -1,4 +1,4 @@
-package com.freshollie.monkeyboard.keystoneradio.activities;
+package com.freshollie.monkeyboard.keystoneradio.ui;
 
 import android.app.Dialog;
 import android.app.DialogFragment;
