+++ app/src/main/res/layout-land/activity_player.xml
@@ -8,7 +8,7 @@
     android:orientation="horizontal"
     android:clipToPadding="false"
     android:clipChildren="false"
-    tools:context=".keystoneradio.activities.PlayerActivity">
+    tools:context=".keystoneradio.ui.PlayerActivity">
     <LinearLayout
         android:layout_width="0dp"
         android:layout_height="match_parent"
