+++ app/src/main/res/layout/activity_player.xml
@@ -8,7 +8,7 @@
     android:layout_height="match_parent"
     android:clipToPadding="false"
     android:orientation="vertical"
-    tools:context=".keystoneradio.activities.PlayerActivity">
+    tools:context=".keystoneradio.ui.PlayerActivity">
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="0dp"
